[fl] chore: display cli command in a span
