+++ libs/filonov/filonov/entrypoints/streamlit_app.py
@@ -20,7 +20,6 @@ import datetime
 import json
 import os
 import tempfile
-from collections.abc import Sequence
 
 import media_fetching
 import media_similarity
@@ -195,7 +194,9 @@ def streamlit_app():
           output_name=name
         ),
       )
-      cli_command = build_cli_command(request, settings.media_tagging_db_url)
+      cli_command = utils.build_cli_command(
+        request, settings.media_tagging_db_url
+      )
       st.code(cli_command, language='bash', wrap_lines=True)
 
     if submitted:
@@ -209,7 +210,9 @@ def streamlit_app():
         ),
       )
       with st.expander('Show CLI command'):
-        cli_command = build_cli_command(request, settings.media_tagging_db_url)
+        cli_command = utils.build_cli_command(
+          request, settings.media_tagging_db_url
+        )
         st.code(cli_command, language='bash', wrap_lines=True)
       if 'fetching_service' not in st.session_state:
         st.session_state['fetching_service'] = (
@@ -243,41 +246,6 @@ def streamlit_app():
         json.dump(generated_map.to_json(), f)
 
       st.success(f'Creative Map saved to {destination}!')
-
-
-def build_cli_command(
-  request: filonov.CreativeMapGenerateRequest, db: str | None
-) -> str:
-  command_template = (
-    'filonov --source {source} \\\n'
-    '\t--media-type {media_type} \\\n'
-    '\t--tagger {tagger} \\\n'
-    '{source_parameters} \\\n'
-    '\t--output-name {output}'
-  )
-  source = request.source.value
-  source_parameters = []
-  for name, value in request.source_parameters.model_dump().items():
-    if not value or name == 'media_type':
-      continue
-    if isinstance(value, str):
-      source_parameters.append(f'\t--{source}.{name}={value}')
-    elif isinstance(value, Sequence):
-      value_concat = ','.join(value)
-      source_parameters.append(f'\t--{source}.{name}={value_concat}')
-
-  source_parameters = ' \\\n'.join(source_parameters)
-  params = {
-    'source': source,
-    'media_type': request.media_type,
-    'tagger': request.tagger,
-    'output': request.output_parameters.output_name,
-    'source_parameters': source_parameters,
-  }
-  non_db_command = command_template.format(**params).strip()
-  if db:
-    return f'{non_db_command} \\\n\t--db-uri {db}'
-  return non_db_command
 
 
 streamlit_app()
