+++ libs/filonov/filonov/entrypoints/utils.py
@@ -16,9 +16,48 @@
 
 """Utils module for filonov entrypoints."""
 
+from collections.abc import Sequence
+
+import filonov
+
 
 def build_creative_map_destination(path: str):
   """Build correct output path."""
   if path_elements := path.split('.')[0:-1]:
     path = '.'.join(path_elements)
   return f'{path}.json'
+
+
+def build_cli_command(
+  request: filonov.CreativeMapGenerateRequest, db: str | None
+) -> str:
+  command_template = (
+    'filonov --source {source} \\\n'
+    '\t--media-type {media_type} \\\n'
+    '\t--tagger {tagger} \\\n'
+    '{source_parameters} \\\n'
+    '\t--output-name {output}'
+  )
+  source = request.source.value
+  source_parameters = []
+  for name, value in request.source_parameters.model_dump().items():
+    if not value or name == 'media_type':
+      continue
+    if isinstance(value, str):
+      source_parameters.append(f'\t--{source}.{name}={value}')
+    elif isinstance(value, Sequence):
+      value_concat = ','.join(value)
+      source_parameters.append(f'\t--{source}.{name}={value_concat}')
+
+  source_parameters = ' \\\n'.join(source_parameters)
+  params = {
+    'source': source,
+    'media_type': request.media_type,
+    'tagger': request.tagger,
+    'output': request.output_parameters.output_name,
+    'source_parameters': source_parameters,
+  }
+  non_db_command = command_template.format(**params).strip()
+  if db:
+    return f'{non_db_command} \\\n\t--db-uri {db}'
+  return non_db_command
